requirements.txt

pytesseract
Flask
Django
django-crispy-forms
Pillow
pandas
numpy
xlrd
requests
lxml
image